+++ app.json
@@ -1,7 +1,7 @@
 {
   "expo": {
     "name": "practice-react-native",
-    "slug": "snack-7d2cf059-2edb-4b06-a246-0924ede885ca",
+    "slug": "snack-96e9475e-eaae-445d-983c-b4c5da5c3c53",
     "version": "1.0.0",
     "orientation": "portrait",
     "icon": "./assets/icon.png",
