+++ screens/ForgotPassword.js
@@ -1,65 +1,25 @@
 // Import Core Libraries
-import { useState, useCallback } from 'react';
-import {
-  View,
-  Text,
-  ScrollView,
-  StyleSheet,
-  SafeAreaView,
-  RefreshControl,
-} from 'react-native';
-
-// Import Components
-import Icon from './components/Icon';
+import { Text, StyleSheet } from 'react-native';
 
 // Import Partials
 import ForgotPasswordForm from './partials/ForgotPasswordForm';
 
-// Import Helpers
-import {
-  horizontalScale,
-  verticalScale,
-  moderateScale,
-} from './helpers/Responsive';
+// Import Layout
+import MainLayout from './layouts/MainLayout';
 
-const wait = (timeout) => {
-  return new Promise((resolve) => setTimeout(resolve, timeout));
-};
+// Import Helpers
+import { verticalScale, moderateScale } from './helpers/Responsive';
 
 export default function Login({ navigation }) {
-  const [refreshing, setRefreshing] = useState(false);
-
-  const onRefresh = useCallback(() => {
-    setRefreshing(true);
-    wait(2000).then(() => setRefreshing(false));
-  }, []);
-
   return (
-    <SafeAreaView style={styles.container}>
-      <ScrollView
-        showsVerticalScrollIndicator={false}
-        refreshControl={
-          <RefreshControl refreshing={refreshing} onRefresh={onRefresh} />
-        }>
-        <View style={styles.body}>
-          <Icon />
-          <Text style={styles.title}>Forgot Password</Text>
-          <ForgotPasswordForm navigation={navigation} />
-        </View>
-      </ScrollView>
-    </SafeAreaView>
+    <MainLayout>
+      <Text style={styles.title}>Forgot Password</Text>
+      <ForgotPasswordForm navigation={navigation} />
+    </MainLayout>
   );
 }
 
 const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    paddingHorizontal: horizontalScale(25),
-  },
-  body: {
-    paddingHorizontal: horizontalScale(25),
-  },
   title: {
     fontFamily: 'Roboto-Medium',
     fontSize: moderateScale(28),
