+++ screens/layouts/MainLayout.js
@@ -1,0 +1,54 @@
+// Import Core Libraries
+import { useState, useCallback } from 'react';
+import {
+  View,
+  ScrollView,
+  StyleSheet,
+  SafeAreaView,
+  RefreshControl,
+} from 'react-native';
+
+// Import Components
+import Icon from '../components/Icon';
+
+// Import Helpers
+import { horizontalScale } from '../helpers/Responsive';
+
+const wait = (timeout) => {
+  return new Promise((resolve) => setTimeout(resolve, timeout));
+};
+
+export default function Login({ children }) {
+  const [refreshing, setRefreshing] = useState(false);
+
+  const onRefresh = useCallback(() => {
+    setRefreshing(true);
+    wait(2000).then(() => setRefreshing(false));
+  }, []);
+
+  return (
+    <SafeAreaView style={styles.container}>
+      <ScrollView
+        showsVerticalScrollIndicator={false}
+        refreshControl={
+          <RefreshControl refreshing={refreshing} onRefresh={onRefresh} />
+        }>
+        <View style={styles.body}>
+          <Icon />
+          {children}
+        </View>
+      </ScrollView>
+    </SafeAreaView>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    paddingHorizontal: horizontalScale(25),
+  },
+  body: {
+    paddingHorizontal: horizontalScale(25),
+  },
+});
